--- routers/api/v1/activity/feeds.go
@@ -1,68 +1,0 @@
-// Copyright 2023 The Gitea Authors. All rights reserved.
-// SPDX-License-Identifier: MIT
-
-package activity
-
-import (
-	"net/http"
-
-	activities_model "code.gitea.io/gitea/models/activities"
-	"code.gitea.io/gitea/modules/context"
-	"code.gitea.io/gitea/routers/api/v1/utils"
-	"code.gitea.io/gitea/services/convert"
-)
-
-func ListUserActivityFeeds(ctx *context.APIContext) {
-	// swagger:operation GET /users/{username}/activities/feeds user userListActivityFeeds
-	// ---
-	// summary: List a user's activity feeds
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: username
-	//   in: path
-	//   description: username of user to get
-	//   type: string
-	//   required: true
-	// - name: only-performed-by
-	//   in: query
-	//   description: if true, only show actions performed by the requested user
-	//   type: boolean
-	// - name: date
-	//   in: query
-	//   description: the date of the activities to be found, format is YYYY-MM-DD
-	//   type: string
-	//   format: date
-	// - name: page
-	//   in: query
-	//   description: page number of results to return (1-based)
-	//   type: integer
-	// - name: limit
-	//   in: query
-	//   description: page size of results
-	//   type: integer
-	// responses:
-	//   "200":
-	//     "$ref": "#/responses/ActivityFeedsList"
-
-	includePrivate := ctx.IsSigned && (ctx.Doer.IsAdmin || ctx.Doer.ID == ctx.ContextUser.ID)
-	listOptions := utils.GetListOptions(ctx)
-
-	opts := activities_model.GetFeedsOptions{
-		RequestedUser:   ctx.ContextUser,
-		Actor:           ctx.Doer,
-		IncludePrivate:  includePrivate,
-		OnlyPerformedBy: ctx.FormBool("only-performed-by"),
-		Date:            ctx.FormString("date"),
-		ListOptions:     listOptions,
-	}
-
-	feeds, count, err := activities_model.GetFeeds(ctx, opts)
-	if err != nil {
-		ctx.Error(http.StatusInternalServerError, "GetFeeds", err)
-		return
-	}
-	ctx.SetTotalCountHeader(count)
-
-	ctx.JSON(http.StatusOK, convert.ToActivities(ctx, feeds, ctx.Doer))
-}
